refac functionalities + definitions testing
