+++ client/testFixture/definition.wlk
@@ -1,0 +1,11 @@
+object pepita {
+  method vola() {
+    
+  }
+}
+
+object manolo {
+  method hacerVolarAPepita() {
+    pepita.vola()
+  }
+}
