+++ server/src/functionalities/symbols.ts
@@ -1,14 +1,30 @@
-import { DocumentSymbol, SymbolKind, WorkspaceSymbol } from 'vscode-languageserver'
+import { DocumentSymbol, DocumentSymbolParams, SymbolKind, WorkspaceSymbol, WorkspaceSymbolParams } from 'vscode-languageserver'
 import { Environment, Field, Method, Module, Node, Package, Program, Test, Variable } from 'wollok-ts'
-import { toVSCRange } from '../utils/text-documents'
+import { packageFromURI, toVSCRange } from '../utils/text-documents'
 import { workspacePackage } from '../utils/vm/wollok'
 
 type Symbolyzable = Program | Test | Module | Variable | Field | Method | Test
 
-export const documentSymbolsFor = (document: Package): DocumentSymbol[] =>
+export const documentSymbols = (environment: Environment) => (params: DocumentSymbolParams): DocumentSymbol[] => {
+  // ToDo this is a temporal fix for https://github.com/uqbar-project/wollok-lsp-ide/issues/61
+  if (!workspacePackage(environment)) {
+    return []
+  }
+  const document = packageFromURI(params.textDocument.uri, environment)
+  if (!document)
+    throw new Error('Could not produce symbols: document not found')
+  return documentSymbolsFor(document)
+}
+
+export const workspaceSymbols = (
+  environment: Environment
+) => (params: WorkspaceSymbolParams): WorkspaceSymbol[] => workspaceSymbolsFor(environment, params.query)
+
+
+const documentSymbolsFor = (document: Package): DocumentSymbol[] =>
   (document.members.filter(isSymbolyzable) as Symbolyzable[]).map(documentSymbol)
 
-export const workspaceSymbolsFor = (environment: Environment, query: string): WorkspaceSymbol[] =>
+const workspaceSymbolsFor = (environment: Environment, query: string): WorkspaceSymbol[] =>
   workspacePackage(environment).descendants.filter(isSymbolyzable)
     .filter(node => node.sourceFileName && node.sourceMap)
     .filter(node => node.name?.toLowerCase().includes(query.toLowerCase()))
